fix arrow + arrayShuffle, added clang-format

+++ XOR_NeuralNetwork/helpers.h
@@ -3,7 +3,7 @@
 
 void free2dArray(double **array, size_t len);
 void arrayFillUniform(double *array, size_t len, double max);
-void arrayShuffle(int *array, size_t len);
+void arrayShuffle(size_t *array, size_t len);
 void print1dArray(double *array, size_t len);
 void print2dArray(double **array, size_t len1, size_t len2);
